requirements.txt
# Pulsara IVR v2 Requirements

# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn>=0.21.1
starlette>=0.26.1
websockets>=11.0.1

# Twilio
twilio>=8.0.0

# HTTP and WebSockets (for direct API calls)
httpx>=0.24.0
websockets>=11.0.1

# Utilities
python-dotenv>=1.0.0
pydantic>=1.10.7
pytz>=2023.3
python-multipart>=0.0.6

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
